# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Vault client"
basePath: "/"
schemes:
- "http"
paths:
  /aws/sns/sts:
    get:
      summary: "Get creds for SNS"
      description: "Prints AWS environment variables for SNS on the console."
      operationId: "getSnsCreds"
      responses:
        204:
          description: "successful operation"
  /aws/sns/{topicArn}:
    post:
      summary: "Publish to SNS"
      description: "Publishes a message to the topic."
      operationId: "postMessageToSns"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "topicArn"
        description: "topic to publish to"
        required: true
        type: string
      - in: "body"
        name: "body"
        description: "the message to publish"
        required: true
        schema:
          type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/PublishResponse"
        400:
          description: "Invalid input"
  /transit/keys:
    get:
      summary: "List keys"
      description: "This endpoint returns a list of keys. Only the key names are returned (not the actual keys themselves)."
      operationId: "getListKeys"
      produces:
      - "application/json"
      responses:
        200:
          description: "Encryption keys"
          schema:
            type: array
            items:
              type: string
  /transit/encrypt/{key}:
    put:
      summary: "Encrypt text"
      description: "This endpoint encrypts the provided plaintext using the named key."
      operationId: "putEncrypt"
      produces:
      - "text/plain"
      consumes:
      - "text/plain"
      parameters:
      - in: path
        description: "Encryption key"
        required: true
        name: key
        type: string
      - in: body
        name: plainText
        description: "Plain text to encrypt"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Successfully encrypted"
          schema:
            type: string
  /transit/decrypt/{key}:
    put:
      summary: "Decrypt text"
      description: "This endpoint decrypts the provided ciphertext using the named key."
      operationId: "putDecrypt"
      produces:
      - "text/plain"
      consumes:
      - "text/plain"
      parameters:
      - in: path
        required: true
        description: "Decryption key"
        name: key
        type: string
      - in: body
        name: cipherText
        description: "Ciphertext to decrypt"
        required: true
        schema:
          type: string
      responses:
        200:
          description: "good"
          schema:
            type: string
definitions:
  PublishRequest:
    type: "object"
    properties:
      topicArn:
        type: string
      message:
        type: string
  PublishResponse:
    type: "object"
    properties:
      messageId:
        type: string